--- a/.travis.yml
+++ b/.travis.yml
@@ -46,7 +46,7 @@ before_install:
   - wget -q "${ANDROID_TOOLS_URL}" -O android-sdk-tools.zip
   - unzip -q android-sdk-tools.zip -d ${ANDROID_HOME}
   - rm android-sdk-tools.zip
-  - mkdir ~/.android
+  - mkdir -p ~/.android
   - echo 'count=0' > ~/.android/repositories.cfg
   - yes | sdkmanager --licenses >/dev/null
   - yes | sdkmanager "platform-tools"