--- a/travisscript/.travis.yml.old
+++ b/travisscript/.travis.yml.old
@@ -1,6 +1,7 @@
-language: bash # Cannot use android for kvm
+language: android
+jdk: oraclejdk8
 sudo: true
-dist: bionic # Bionic is needed for KVM Nested Virtualization
+dist: trusty # Use Ubuntu 14.04 as its faster
 before_cache:
   - rm -f  $HOME/.gradle/caches/modules-2/modules-2.lock
   - rm -fr $HOME/.gradle/caches/*/plugin-resolution/
@@ -14,20 +15,13 @@ cache:
 env:
   global:
     - API_LEVEL=28
-    - ANDROID_EMULATOR_API_LEVEL=28 # Try on 28 first for emulator, if it works we can go 29
+    - ANDROID_EMULATOR_API_LEVEL=25
     - ANDROID_BUILD_TOOLS_VERSION=28.0.3
-    - ANDROID_ABI=x86_64
-    - ANDROID_FLAVOUR=google_apis # Use default for non-google
+    - ANDROID_ABI=google_apis/arm64-v8a
     - ADB_INSTALL_TIMEOUT=20
     - ANDROID_TARGET=android-25
     - ANDROID_TAG=google_apis
     - DISCORD_URL=https://raw.githubusercontent.com/itachi1706/travis-ci-discord-webhook/master/send.sh
-    # KVM Self setup of Android SDK
-    - ANDROID_HOME=${HOME}/android-sdk # Default location for SDK
-    - ANDROID_TOOLS_URL="https://dl.google.com/android/repository/sdk-tools-linux-4333796.zip" # From Android Dev Site
-    - JDK="1.8" # Android JDK
-    - TOOLS=${ANDROID_HOME}/tools # SDK Tools Location
-    - PATH=${ANDROID_HOME}:${ANDROID_HOME}/emulator:${TOOLS}:${TOOLS}/bin:${ANDROID_HOME}/platform-tools:${PATH}
   matrix:
     - ESPRESSO=1 # Espresso Instrumentation Tests
     - ESPRESSO=0 # Unit Tests
@@ -35,55 +29,29 @@ matrix:
     allow_failures:
         - env: ESPRESSO=1
     fast_finish: true
+android:
+  components:
+    - tools
+    - platform-tools
+    - tools
+    - build-tools-$ANDROID_BUILD_TOOLS_VERSION
+    - android-$API_LEVEL
+    - android-$ANDROID_EMULATOR_API_LEVEL
+    - extra-google-m2repository
+    - extra-android-m2repository
+    - addon-google_apis-google-$ANDROID_EMULATOR_API_LEVEL
+    - sys-img-arm64-v8a-google_apis-$ANDROID_EMULATOR_API_LEVEL
 before_install:
-  # Setup Android SDK JDK
-  - export TARGET_JDK="${JDK}"
-  - JDK="1.8"
-  - chmod +x ./travisscript/install-jdk-travis.sh
-  - source ./travisscript/install-jdk-travis.sh
-
-  # Setup Android SDK
-  - wget -q "${ANDROID_TOOLS_URL}" -O android-sdk-tools.zip
-  - unzip -q android-sdk-tools.zip -d ${ANDROID_HOME}
-  - rm android-sdk-tools.zip
-  - mkdir ~/.android
-  - echo 'count=0' > ~/.android/repositories.cfg
-  - yes | sdkmanager --licenses >/dev/null
-  - yes | sdkmanager "platform-tools"
-  - yes | sdkmanager "tools"
-  - yes | sdkmanager "build-tools;$ANDROID_BUILD_TOOLS_VERSION"
-  - yes | sdkmanager "platform;android-$API_LEVEL" # Current Android compileSDK
-install:
-  # Download Emulator Files
-  - yes | sdkmanager "platform;android-$ANDROID_EMULATOR_API_LEVEL" # Android Emulator Platform
-  - yes | sdkmanager "emulator"
-  - yes | sdkmanager "system-images;android-$ANDROID_EMULATOR_API_LEVEL;$ANDROID_FLAVOUR;$ANDROID_ABI" # Android Emulator Image
-
-  # Setup Linux KVM
-  - sudo -E apt-get -yq --no-install-suggests --no-install-recommends install bridge-utils libpulse0 libvirt-bin qemu-kvm virtinst ubuntu-vm-builder
-  - sudo adduser $USER libvirt
-  - sudo adduser $USER kvm
-
-  # Launch Android Emulator
-  - android list targets
-  - chmod +x ./travisscript/startemu.sh
-  - ./travisscript/startemu.sh
-
-  # Switch back to target JDK
-  - JDK="${TARGET_JDK}"
-  - source ~/.install-jdk-travis.sh
-before_script:
-  # Original Steps continued
-  #- mkdir "$ANDROID_HOME/licenses" || true
-  #- echo -e "\n8933bad161af4178b1185d1a37fbf41ea5269c55" > "$ANDROID_HOME/licenses/android-sdk-license"
-  #- echo -e "\n84831b9409646a918e30573bab4c9c91346d8abd" > "$ANDROID_HOME/licenses/android-sdk-preview-license"
-  #- yes | sdkmanager "platforms;android-28" # Temporary fix for checksum mismatch (travis-ci/#8874)
+  - mkdir "$ANDROID_HOME/licenses" || true
+  - echo -e "\n8933bad161af4178b1185d1a37fbf41ea5269c55" > "$ANDROID_HOME/licenses/android-sdk-license"
+  - echo -e "\n84831b9409646a918e30573bab4c9c91346d8abd" > "$ANDROID_HOME/licenses/android-sdk-preview-license"
+  - yes | sdkmanager "platforms;android-28" # Temporary fix for checksum mismatch (travis-ci/#8874)
   - openssl aes-256-cbc -K $encrypted_2969307dece5_key -iv $encrypted_2969307dece5_iv -in secrets.tar.enc -out secrets.tar -d
   - tar xvf secrets.tar
-  #- android list targets
+  - android list targets
   - chmod +x gradlew
-  #- chmod +x ./travisscript/startemu.sh
-  #- ./travisscript/startemu.sh
+  - chmod +x ./travisscript/startemu.sh
+  - ./travisscript/startemu.sh
   #- ./gradlew clean # Uncomment to clean cache
   - travis_retry ./gradlew assembleDebug
   - travis_retry ./gradlew assembleDebugAndroidTest