This PR finally allows us to make use of Android x86_64 emulators and bump our emulator to API 29 (Android 10)

This is due to the fact that we have converted the Travis CI build to allow us to make use of KVM Virtualization for hardware acceleration for emulators.

We can also clearly see the difference in terms of build speeds right now as well

**Old Build Times**   

| Build Mode | Timings |
| --- | --- |
| Non-Espresso | 12 minutes 38 seconds |
| Espresso | 26 minutes 13 seconds |
| Total | 38 minutes 51 seconds |

**New Build Times**

Build Mode | Timings | Faster by (%) (3sf)
------------ | ------------- | ----
Non-Espresso | 7 minutes 41 seconds | 48.7%
Espresso | 17 minutes 10 seconds | 41.7%
Total | 24 minutes 51 seconds | 44.0%

Resolves CUTILAND-363